Add files via upload
20240825 build

+++ Iconinator.py
@@ -1,27 +1,25 @@
 #!/env/Python3.10.4
 #/MobCat (2024)
 
+# This tool will download your UIX Lite config and default.xip
+# read them, download icons from your xbox UData 
+
 #WARNING:
-# This is just a PoC
-# It was built in an afternoon, it's very sloppy
-# but it sorta works, but still needs some work to be done.
-# And even then, I hope this "tool" gets outdated and redundant
-# as soon as the xbox or UIX can build this icon list on it's own.
+# This tool is """beta"""
+# It works, but it's sloppy and needs work of it's own.
 
 #TODO:
-# Figure out the xbx xip packing thing.
-# Replace XBEJson with pyxbe when ready.
-#TODO QoL:
-# add yes to all for dbcheck
+# Try and rip icons from xbes first, then download from UData, and if both of those fail, then finally try an icon CDN download.
 
 import configparser
 import ftplib
-import requests
+from requests import get # get icons from CDN
 import os
-import socket                    # Probs dont need the whole sockets lib...
-import sys                       # Silly little sys.exit()
-import json                      # Remove this lib once pyxbe is working again.
-import subprocess                # Remove this lib once pyxbe is working again.
+from socket import socket, AF_INET, SOCK_DGRAM # Get users ip / network ip prefix.
+import sys        # Silly little sys.exit()
+import json       # Remove this lib once pyxbe is working again and we can remove XBEJson.exe.
+import subprocess # Needed for xip.exe and XBEJson.exe
+import shutil
 
 # temp removed, lib has some bugs, it's okie.
 #from xbe import Xbe 
@@ -36,13 +34,15 @@ print('''
 ▒▒▒ ▒▒▒▒▒ ▒▒▒▒▒▒   ▒▒▒▒▒▒ ▒▒▒▒▒ ▒▒▒   ▒▒   ▒▒▒   ▒▒  ▒▒   ▒▒▒ ▒▒ ▒▒  ▒▒▒  ▒▒  ▒▒▒▒▒▒
 ░░░░     ░░░░         ░░   ░░░   ░░   ░░   ░░░   ░░░    ░ ░░░░  ░░░░░   ░░░░  ░░░░░░
 ────────────────────────────────────────────────────────────────────────────────────
-Iconinator 20240823
+Iconinator 20240825
+Automatically building an icon pack for the games launcher,
+based on the content you have installed on your Xbox HDD.
 ''')
 
 # Function to get users IP
-# Just a simple cheat so you dont have to type a full ip addess.
+# Just a simple cheat so you don't have to type a full ip address.
 def getIP():
-	s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+	s = socket(AF_INET, SOCK_DGRAM)
 	try:
 		# doesn't even have to be reachable
 		s.connect(('10.255.255.255', 1))
@@ -51,8 +51,7 @@ def getIP():
 		IP = '127.0.0.1'
 	finally:
 		s.close()
-	return '.'.join(IP.split('.')[:3]) # Weid hack but ok.
-
+	return '.'.join(IP.split('.')[:3]) # Weird hack but ok.
 
 
 # Load custom or default settings ini based on command args.
@@ -64,9 +63,9 @@ else:
 	settingsINI = sys.argv[1]
 	print(f"Loading {settingsINI}")
 
-# Build a new config if the config we are trying to load does not excist.
+# Build a new config if the config we are trying to load does not exist.
 if not os.path.isfile(settingsINI):
-	print(f"{settingsINI} Does not excist, lets make one.")
+	print(f"{settingsINI} Does not exist, lets make one.")
 	ipPrefix = getIP()
 	iniXboxIP = input(f"Enter your xboxs IP address: {ipPrefix}.")
 	try:
@@ -76,9 +75,7 @@ if not os.path.isfile(settingsINI):
 		sys.exit()
 
 	print(f"Saving {settingsINI} with default ftp login creds.\n")
-	#config['UIXinator'] = {'xboxIP': iniXboxIP, 'ftpLogin': 'xbox:xbox'}
-	jsonConfig = {}
-	config['UIXinator'] = {'xbox_IP': f'{ipPrefix}.{iniXboxIP}', 'ftp_Login': 'xbox:xbox'}
+	config['UIXinator'] = {'xbox_IP': f'{ipPrefix}.{iniXboxIP}', 'ftp_Login': 'xbox:xbox', 'cleanup': 'True', 'auto_reboot': 'False', 'iconCDN': 'https://raw.githubusercontent.com/MobCat/MobCats-original-xbox-game-list/main/xbx'}
 	with open(settingsINI, 'w') as newConfig:
 		config.write(newConfig)
 
@@ -86,8 +83,10 @@ if not os.path.isfile(settingsINI):
 config.read(settingsINI)
 defultCreds = config.get('UIXinator', 'ftp_Login')
 ftpUserPass = defultCreds.split(':')
-xboxIP = config.get('UIXinator', 'xbox_IP')
-
+xboxIP      = config.get('UIXinator', 'xbox_IP')
+CleanupFlag = config.getboolean('UIXinator', 'cleanup')
+iconCDN     = config.get('UIXinator', 'iconCDN')
+rebootUIX   = config.getboolean('UIXinator', 'auto_reboot')
 
 # Basically just check and setup our ftp connection.
 print(f"Connecting to {defultCreds}@{xboxIP}")
@@ -119,6 +118,7 @@ def DirLST(ftp_obj, path="."):
 	return templst
 
 def DownloadFTP(ftp_obj, remote_path, local_path):
+	ftp_obj.cwd("/")
 	try:
 		ftp_obj.size(remote_path)
 		with open(local_path, "wb") as f:
@@ -127,8 +127,11 @@ def DownloadFTP(ftp_obj, remote_path, local_path):
 		return f"FTP Error {remote_path}: {e}"
 
 def DownloadWEB(titleID):
-	r = requests.get(f'https://raw.githubusercontent.com/MobCat/MobCats-original-xbox-game-list/main/xbx/{titleID[:4]}/{titleID}.xbx', allow_redirects=True)
-	open(f'xbx/{titleID}.xbx', 'wb').write(r.content)
+	r = get(f'{iconCDN}/{titleID[:4].upper()}/{titleID.upper()}.xbx', allow_redirects=True)
+	if r.status_code == 200:
+		open(f'xbx/{titleID}.xbx', 'wb').write(r.content)
+	else:
+		return f'ERROR: {r.status_code}'
 
 # Load UIX Config
 err = DownloadFTP(ftp, f'C/UIX Configs/config.xbx', f'config.xbx')
@@ -138,21 +141,22 @@ if err is None:
 	UIXconfig.read('config.xbx')
 	cofigPaths = []
 	for option in UIXconfig.options('LauncherMenu'):
-		if option.startswith('path') and UIXconfig.get('LauncherMenu', option): # TXT is Uppercase, ConfigParser reades it as lowercase??
+		if option.startswith('path') and UIXconfig.get('LauncherMenu', option): # TXT is Uppercase, ConfigParser reads it as lowercase??
 			path = UIXconfig.get('LauncherMenu', option)
 			cofigPaths.extend(path.split(';'))
 
 	# Remove dupes
 	cofigPaths = list(set(cofigPaths))
 	# Cleanup
-	os.remove('config.xbx')
+	if CleanupFlag:
+		os.remove('config.xbx')
 
-	# Janky build a list of all dash paths for all partisions, E/Games, F/Games, etc..
+	# Janky build a list of all dash paths for all partitions, E/Games, F/Games, etc..
 	ftp.cwd("/")
 	contents = ftp.nlst()
 	systemBlacklist = ['C', 'X', 'Y', 'Z']
 	partishionsLst = [item[-1:] for item in contents if item[-1:] not in systemBlacklist]
-	# Now build our finaly list of all dashPaths
+	# Now build our finally list of all dashPaths
 	dashPaths = []
 	for partition in partishionsLst:
 		for path in cofigPaths:
@@ -166,8 +170,17 @@ else:
 	print(f"{err}\nIs UIX Lite installed onto C? I can't seem to find your config.xbx")
 	sys.exit()
 
+print("Downloading your icons xip to prepare it for updating...")
+err = DownloadFTP(ftp, f'C/xboxdashdata.185ead00/default.xip', f'default.xip')
+if err is not None:
+	print(err)
+	sys.exit()
 
 ###########################################
+# Check if we have a place to save icons too.
+if not os.path.exists('xbx'):
+    os.makedirs('xbx')
+
 UIXpaths = []
 titleIDs = [] # Use for later
 print("Downloading xbes and extracting title IDs...\nPlease wait, some large xbes will take time to download...")
@@ -177,7 +190,7 @@ for iniDir in dashPaths:
 		maxlen = len(dirlst)
 		cnt = 1
 		for gameDir in dirlst:
-			print(f"[{iniDir} {cnt}/{maxlen}] Downloading xbes and extracting title IDs...       ", end="\r")
+			print(f"[{iniDir} {cnt}/{maxlen}] Downloading xbes and extracting title IDs...              ", end="\r")
 			DownloadFTP(ftp, f'{gameDir}/default.xbe', 'default.xbe')
 
 			# pyxbe almost worked, but it's having virtual address offset issues..
@@ -186,7 +199,7 @@ for iniDir in dashPaths:
 			#titleID = hex(xbe.cert.title_id)[2:]
 
 			# Backup, Use XBEJson to dump xbe info as json and read back said json.
-			conout = subprocess.run(f'XBEJson.exe default.xbe', stdout=subprocess.PIPE).stdout.decode()
+			conout = subprocess.run(f'lib/XBEJson.exe default.xbe', stdout=subprocess.PIPE).stdout.decode()
 			xbeJson = json.loads(conout)
 			titleID = xbeJson['Title_ID'].lower()
 			UIXpaths.append(f'{gameDir.split("/")[-1]}={titleID}')
@@ -194,14 +207,14 @@ for iniDir in dashPaths:
 			cnt += 1
 		cnt = 1
 	else:
-		#input(f"{iniDir} Does not excist on your xbox.")
-		print(f"[{iniDir} ?/?] This file path does not exist on your xbox.                      ", end="\r")
+		#input(f"{iniDir} Does not exist on your xbox.")
+		print(f"[{iniDir} ?/?] This file path does not exist on your xbox.                              ", end="\r")
 
 # quick house keeping.
 os.remove('default.xbe')
 
 
-print(f"\nFound {len(titleIDs)} titles installed on your xbox currently.\nSaveing and uploading UIX icons config...")
+print(f"\nFound {len(titleIDs)} titles installed on your xbox currently.\nSaving and uploading UIX icons config...")
 with open('Icons.xbx', 'w') as f:
 	f.write(f"[default]\n") # I think if we change this to something like [icons] we can load icons form a icon.xip not default.xip maybe..
 	for line in UIXpaths:
@@ -213,117 +226,54 @@ with open('Icons.xbx', "rb") as f:
 	ftp.storbinary("STOR Icons.xbx", f)
 
 
-# Check to see if you have a some custom icons installed
-#BUG: This check will fail if UIX Lite is not installed, *However* the download config and sys exit stage *should* cover it.. should..
-#BUGBUG: If default.xip is *missing* from your xbox, this script will just crash. we should do a try: or smh, but tbh if default.xip is missing you have bigger issues that launcher icons can't solve.
-NewDownload = False
-ftp.cwd("/")
-xipSize = ftp.size('C/xboxdashdata.185ead00/default.xip')
-if xipSize <= 1081701:
-	GetAIO = input(f'''
-It appears you have the stock icon pack installed, or not a lot of icons installed.
-Would you like to D̲ownload a pre-made icon pack or
-Build a C̲ustom icon pack from content installed on your xbox.
+#####################################################
 
-(PLEASE NOTE: In this beta only downloading custom icons is implemented. Just prese Enter.
-idk how to build my own xip file yet.)
-Input selection [D/c]: ''')
-	if GetAIO != 'c':
-		print("Downloading custom icon pack...")
-		#TODO: Check if we already have it downloaded so we dont overwright it.
-		#TODO: Change this link to the UIX github repo or somewhere more "officle" then my repo.
-		url = 'https://raw.githubusercontent.com/MobCat/UIXinator/main/default.xip'
-		response = requests.get(url, stream=True)
-		total_size = int(response.headers.get('content-length', 0))
-		progress_bar_size = 50
-		with open('default.xip', 'wb') as f:
-			for data in response.iter_content(1024): #1 Kibibyte chunk size
-				f.write(data)
-				progress = int(f.tell() * progress_bar_size / total_size)
-				print(f"\r[{'█' * progress}{' ' * (progress_bar_size - progress)}] {f.tell()/1024/1024:.2f} MB / {f.tell() * 100 / total_size:.2f}% ", end="")
-				
-		print("\nDownload complete.\nUploading default.xip to your xbox..")
-		with open('default.xip', "rb") as f:
-			ftp.cwd("/")
-			ftp.cwd("C/xboxdashdata.185ead00/")
-			ftp.storbinary("STOR default.xip", f)
-		NewDownload = True
-
-#Kinda a weird check but I want to skip asking the user to reupload something they JUST downloaded
-#It felt awkward
-if os.path.exists('default.xip') and NewDownload != True:
-	reupCheck = input("\nIt appears you already have a default.xip ready to go,\nWould you like to re-upload it and install it to your xbox? [Y/n]: ")
-	if reupCheck != 'n':
-		print("Uploading default.xip to your xbox..")
-		with open('default.xip', "rb") as f:
-			ftp.cwd("/")
-			ftp.cwd("C/xboxdashdata.185ead00/")
-			ftp.storbinary("STOR default.xip", f)
-
-print("All Done \\^__^/")
-print("\nEnd of Beta\nMore to come soon™")
-input("Press Enter to exit.")
-sys.exit()
-
-##################################################################################################################################
-# "Dead" code line
-# It's not dead, just not ready yet. but it works. so I'm not removing it, I'm just waiting for external tools and stuffs.
-##################################################################################################################################
-
-# Rip icons from xbox or download new ones.
-try:
-	iconSelect = input("""
-[Icon downloader]
-1. Download Icons from your console
-   (Fast, however you need to run the game first, before you can download it's icon)
-
-2. Download from online database
-   (Slow and incomplete, but you don't need to run the game first)
-   (Database only contains icons for 'Retail Games')
-
-Press Ctrl+c to quit
-Your Icons.xbx file was built and sent to your console but you don't need any new icons
-
-Default input is 1.
-Select input: """)
-except KeyboardInterrupt:
-	print("Goodbye ^__^/")
-	sys.exit()
-
-# Setup a place to download icons to.
-if not os.path.exists('xbx'):
-	os.makedirs('xbx')
-
-if iconSelect == "2":
-	print("Downloading...")
-	maxlen = len(titleIDs)
-	cnt = 1
-	for titleID in titleIDs:
-		print(f"[{cnt}/{maxlen}] {titleID}.xbx", end="\r")
-		DownloadWEB(titleID.upper())
-		cnt += 1
-else:
-	for titleID in titleIDs:
-		ftp.cwd("/")
-		err = DownloadFTP(ftp, f'E/UData/{titleID}/TitleImage.xbx', f'xbx/{titleID}.xbx')
+# Rip icons from xbox, if the icon is not in UData, try downloading one.
+#BUG: Not all icons are on CND, homebrew uses whatever title IDs they want.
+# We are not ripping icons for xbes yet, this may solve some of these issues, and cause others.
+errcnt = 0
+for titleID in titleIDs:
+	ftp.cwd("/")
+	err = DownloadFTP(ftp, f'E/UData/{titleID}/TitleImage.xbx', f'xbx/{titleID}.xbx')
+	if err is not None:
+		print(f'{titleID}.xbx Not found on xbox, Downloading from db...         ', end="\r")
+		err = DownloadWEB(titleID)
 		if err is not None:
-			print(err)
-			dbcheck = input("Try and grab it from the online database? [Y/n]: ")
-			if dbcheck != "n":
-				print(f'Downloading {titleID}.xbx')
-				DownloadWEB(titleID)
+			print(f'{titleID}.xbx Not on CDN...                                 ', end="\r")
+			errcnt += 1
 
+if errcnt > 0:
+	print(f"""
+{errcnt} icons where missing from the CDN
+This is normal for homebrew apps.
+Please run the game or homebrew first
+then this tool can grab the icon from the games save data.""")
 
-#TODO:
-# Now we have a folder of xbx icons based on the games in your Icons.xbx list,
-# we need to some how pack them into the default.xip
-# idk how to do that...
+print("\nUpdating your icons xip...")
+subprocess.call("lib/xip.exe -m -c default.xip xbx/*.xbx", stdout=subprocess.DEVNULL)
+
+with open('default.xip', "rb") as f:
+	ftp.cwd("/")
+	ftp.cwd("C/xboxdashdata.185ead00/")
+	ftp.storbinary("STOR default.xip", f)
+
+if CleanupFlag:
+	os.remove('default.xip')
+	shutil.rmtree("xbx")
+
+if rebootUIX:
+	print("Hot rebooting to UIX Lite.")
+	ftp.sendcmd('SITE EXEC /C/xboxdash.xbe')
+
+ftp.quit()
+
+input("All done \\^__^/\n Your UIX Lite launcher icons have been updated.\nPress Enter to exit.")
 
 #Note:
 # When/if we try and build custom xip from users installed games
 # Check if the default.xbe is 6924 bytes in side, we assume this is an xiso attacher and wont have an icon
 # So we need to rip it from the save data or download one from the title id database.
-# If we dont wanna play it fast'n'lose with basic file size checking
+# If we don't wanna play it fast'n'lose with basic file size checking
 # We could load and check the raw data of the xbe
 # 0x00001AA0 to 0x00001AAD == '\Device\CdRom1'
 # 0x00001AF8 to 0x00001AFF == '!"# .iso'
